move data logic to separate hook
